tam_audio_matching/requirements.txt
pydub
python-dotenv
librosa
cachetools
opencv-python
configparser
psycopg2-binary
pysftp
matplotlib
scikit-image
noisereduce